make small miscellaneous changes

+++ search.c
@@ -198,7 +198,7 @@ static void moonfish_expand(struct moonfish_node *node, struct moonfish_chess *c
 	node->count = child_count;
 }
 
-static float moonfish_confidence(struct moonfish_node *node)
+static double moonfish_confidence(struct moonfish_node *node)
 {
 	if (node->visits == 0) return 1e9;
 	return 1 / (1 + pow(10, node->score / 400.0)) + 2 * sqrt(log(node->parent->visits) / node->visits);
@@ -221,7 +221,7 @@ static void moonfish_node_chess(struct moonfish_node *node, struct moonfish_ches
 static struct moonfish_node *moonfish_select(struct moonfish_node *node, struct moonfish_chess *chess)
 {
 	struct moonfish_node *next;
-	float max_confidence, confidence;
+	double max_confidence, confidence;
 	int i, count;
 	
 	for (;;) {
